+++ extensions/zen-browser/src/hooks/usePinnedTabs.ts
@@ -1,0 +1,6 @@
+import { usePromise } from "@raycast/utils";
+import { readBridgeTabs } from "../util/bridge";
+
+export function usePinnedTabs() {
+  return usePromise(readBridgeTabs, [], { onError: () => {} });
+}